requirements.txt
coloredlogs==15.0.1
easydict==1.10
einops==0.7.0
glumpy==1.1.0
imageio==2.5.0
ipdb==0.13.13
ipython==8.12.2
matplotlib==2.2.4
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==9.4.0
Pillow==10.1.0
positional_encodings==6.0.1
pycocotools==2.0
PyOpenGL==3.1.0
PyOpenGL==3.1.7
pypng==0.0.19
pytz==2023.3.post1
scikit_learn==1.3.0
scipy==1.11.4
setuptools==68.0.0
skimage==0.0
timm==0.9.7
torchvision==0.12.0
tqdm==4.66.1
vispy==0.14.0
webdataset==0.2.57
